feat : 전체 로직에 토큰 유효성 검사 추가, 전체 도서 조회 페이지네이션 추가
1) userId가 필요한 로직의 경우 토큰에서 가져올 수 있도록 수정
2) 로그인 해야 동작할 수 있는 기능을 구분해서 토큰 검사 추가
3) 전체 도서 조회 페이지네이션 리턴하도록 수정

+++ week_8-1/controller/CartController.js
@@ -1,12 +1,11 @@
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
+const ensureAuthorization = require("../middleware/auth");
 const jwt = require('jsonwebtoken');
-const dotenv = require('dotenv');
-dotenv.config();
 
 const addToCart = (req, res) => {
     const {bookId, quantity} = req.body;
-    const authorization = ensureAuthorization(req, res);
+    const authorization = ensureAuthorization(req);
 
     if(authorization instanceof jwt.TokenExpiredError) {
         return res.status(StatusCodes.UNAUTHORIZED).json({
@@ -34,7 +33,7 @@ const addToCart = (req, res) => {
 const getCartItems = (req, res) => {
     const {selected} = req.body;
     // selected = [1, 3] 배열의 형태
-    const authorization = ensureAuthorization(req, res);
+    const authorization = ensureAuthorization(req);
 
     // instanceof 좌측이 우측과 같은 모양이니?
     if(authorization instanceof jwt.TokenExpiredError) {
@@ -69,6 +68,22 @@ const getCartItems = (req, res) => {
 }
 
 const removeCartItem = (req, res) => {
+    // userId가 필요한 건 아니지만,
+    // 로그인을 해야만 사용할 수 있는 기능이기 때문에 토큰 체크
+    const authorization = ensureAuthorization(req);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }
+
+    if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }
+
     const {id} = req.params;
     const sql = `DELETE FROM cartItems WHERE id = ?`;
     conn.query(sql, id, function(err, results) {
@@ -78,18 +93,6 @@ const removeCartItem = (req, res) => {
         }
         return res.status(StatusCodes.OK).json(results);
     });
-}
-
-function ensureAuthorization(req, res) {
-    try {
-        const receivedJwt = req.headers["authorization"];
-        const decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-        return decodedJwt;
-    } catch(err) {
-        console.log(err.name);
-        console.log(err.message);
-        return err;
-    }
 }
 
 module.exports = {
